pyproject.toml
[project]
name = "fastapi-mongo-app"
version = "0.1.0"
description = "FastAPI + MongoDB REST API with User CRUD"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.5.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.1.0",
    "aiosmtplib>=3.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-pinecone>=0.2.0",
    "langgraph>=0.2.0",
    "openai>=1.40.0",
    "pinecone-client>=4.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.7.0",
    "python-multipart>=0.0.9",
    "ddgs>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
    "httpx>=0.27.0",
]
